Move B and T to (new) testing.js
